Issue #25:
Refactored flamed3_node to construct the tree from the log entries, rather than a flamegraph string.
Cron manager now makes a check for every sample.
Samples are recorded while inside a task, and this is passed to flamed3_node when constructing the tree.
Created task_samples class to help track samples for a task.
Changed the lower period limit for sampling to 20ms.